+++ .local/share/gnome-shell/extensions/blur-my-shell@aunetx/paint_signals.js
@@ -15,10 +15,18 @@ var PaintSignals = class PaintSignals {
             actor: actor,
             paint_effect: paint_effect
         };
+        let counter = 0;
 
         actor.add_effect(paint_effect);
         this.connections.connect(paint_effect, 'update-blur', () => {
-            blur_effect.queue_repaint();
+            try {
+                // checking if blur_effect.queue_repaint() has been recently called
+                if (counter === 0) {
+                    counter = 2;
+                    blur_effect.queue_repaint();
+                }
+                else counter--;
+            } catch (e) { }
         });
 
         this.buffer.push(infos)