+++ .local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/lib/API.js
@@ -2,14 +2,17 @@
  * API Library
  *
  * @author     Javad Rahmatzadeh <j.rahmatzadeh@gmail.com>
- * @copyright  2020-2021
- * @license    GNU General Public License v3.0
+ * @copyright  2020-2022
+ * @license    GPL-3.0-only
  */
 
 const NOTIFICATION_BANNER_POSITION = {
     TOP_START: 0,
     TOP_CENTER: 1,
     TOP_END: 2,
+    BOTTOM_START: 3,
+    BOTTOM_CENTER: 4,
+    BOTTOM_END: 5,
 };
 
 const PANEL_POSITION = {
@@ -49,7 +52,7 @@ var API = class
     /**
      * Class Constructor
      *
-     * @param {Object} dependecies
+     * @param {Object} dependencies
      *   'Main' reference to ui::main
      *   'BackgroundMenu' reference to ui::backgroundMenu
      *   'OverviewControls' reference to ui::overviewControls
@@ -63,6 +66,8 @@ var API = class
      *   'WindowPreview' reference to ui::windowPreview
      *   'Workspace' reference to ui::workspace
      *   'LookingGlass' reference to ui::lookingGlass
+     *   'MessageTray' reference to ui::messageTray
+     *   'OSDWindow' reference to ui::osdTray
      *   'St' reference to St
      *   'Gio' reference to Gio
      *   'GLib' reference to GLib
@@ -72,34 +77,37 @@ var API = class
      *   'GObject' reference to GObject
      * @param {number} shellVersion float in major.minor format
      */
-    constructor(dependecies, shellVersion)
+    constructor(dependencies, shellVersion)
     {
-        this._main = dependecies['Main'] || null;
-        this._backgroundMenu = dependecies['BackgroundMenu'] || null;
-        this._overviewControls = dependecies['OverviewControls'] || null;
-        this._workspaceSwitcherPopup = dependecies['WorkspaceSwitcherPopup'] || null;
-        this._interfaceSettings = dependecies['InterfaceSettings'] || null;
-        this._searchController = dependecies['SearchController'] || null;
-        this._viewSelector = dependecies['ViewSelector'] || null;
-        this._workspaceThumbnail = dependecies['WorkspaceThumbnail'] || null;
-        this._workspacesView = dependecies['WorkspacesView'] || null;
-        this._panel = dependecies['Panel'] || null;
-        this._windowPreview = dependecies['WindowPreview'] || null;
-        this._workspace = dependecies['Workspace'] || null;
-        this._lookingGlass = dependecies['LookingGlass'] || null;
-        this._st = dependecies['St'] || null;
-        this._gio = dependecies['Gio'] || null;
-        this._glib = dependecies['GLib'] || null;
-        this._clutter = dependecies['Clutter'] || null;
-        this._util = dependecies['Util'] || null;
-        this._meta = dependecies['Meta'] || null;
-        this._gobject = dependecies['GObject'] || null;
+        this._main = dependencies['Main'] || null;
+        this._backgroundMenu = dependencies['BackgroundMenu'] || null;
+        this._overviewControls = dependencies['OverviewControls'] || null;
+        this._workspaceSwitcherPopup = dependencies['WorkspaceSwitcherPopup'] || null;
+        this._interfaceSettings = dependencies['InterfaceSettings'] || null;
+        this._searchController = dependencies['SearchController'] || null;
+        this._viewSelector = dependencies['ViewSelector'] || null;
+        this._workspaceThumbnail = dependencies['WorkspaceThumbnail'] || null;
+        this._workspacesView = dependencies['WorkspacesView'] || null;
+        this._panel = dependencies['Panel'] || null;
+        this._windowPreview = dependencies['WindowPreview'] || null;
+        this._workspace = dependencies['Workspace'] || null;
+        this._lookingGlass = dependencies['LookingGlass'] || null;
+        this._messageTray = dependencies['MessageTray'] || null;
+        this._osdWindow = dependencies['OSDWindow'] || null;
+        this._st = dependencies['St'] || null;
+        this._gio = dependencies['Gio'] || null;
+        this._glib = dependencies['GLib'] || null;
+        this._clutter = dependencies['Clutter'] || null;
+        this._util = dependencies['Util'] || null;
+        this._meta = dependencies['Meta'] || null;
+        this._gobject = dependencies['GObject'] || null;
 
         this._shellVersion = shellVersion;
         this._originals = {};
+        this._timeoutIds = {};
 
         /**
-         * whether seach entry is visible
+         * whether search entry is visible
          *
          * @member {boolean}
          */
@@ -135,6 +143,25 @@ var API = class
     {
         this.UIStyleClassRemove(this._getAPIClassname('shell-version'));
         this._startSearchSignal(false);
+        
+        for (let [name, id] of Object.entries(this._timeoutIds)) {
+            this._glib.source_remove(id);
+            delete(this._timeoutIds[name]);
+        }
+    }
+
+    /**
+     * add to animation duration
+     *
+     * @param {number} duration in milliseconds
+     *
+     * @returns {number}
+     */
+    _addToAnimationDuration(duration)
+    {
+        let settings = this._st.Settings.get();
+
+        return (settings.enable_animations) ? settings.slow_down_factor * duration : 1;
     }
 
     /**
@@ -165,6 +192,9 @@ var API = class
      *  no-window-close
      *  refresh-styles
      *  no-ripple-box
+     *  no-weather
+     *  no-world-clocks
+     *  panel-icon-size
      *
      * @returns {string}
      */
@@ -197,6 +227,9 @@ var API = class
             'no-window-close',
             'refresh-styles',
             'no-ripple-box',
+            'no-weather',
+            'no-world-clocks',
+            'panel-icon-size',
         ];
 
         if (!possibleTypes.includes(type)) {
@@ -218,6 +251,10 @@ var API = class
      */
     panelCornerSetDefault()
     {
+        if (this._shellVersion >= 42) {
+            return;
+        }
+
         let classnameStarter = this._getAPIClassname('panel-corner');
 
         for (let size = 0; size <= 60; size++) {
@@ -234,6 +271,10 @@ var API = class
      */
     panelCornerSetSize(size)
     {
+        if (this._shellVersion >= 42) {
+            return;
+        }
+
         this.panelCornerSetDefault();
 
         if (size > 60 || size < 0) {
@@ -308,8 +349,8 @@ var API = class
 
     /**
      * emit refresh styles
-     * this is useful when an style doesn't emit change because doesn't have
-     * standard styles. for example an style with only `-natural-hpadding`
+     * this is useful when changed style doesn't emit change because doesn't have
+     * standard styles. for example, style with only `-natural-hpadding`
      * won't notify any change. so you need to call this function
      * to refresh that
      *
@@ -334,33 +375,61 @@ var API = class
      */
     _emitPanelPositionChanged(calledFromChanger = false)
     {
-        if (!calledFromChanger) {
-            this.panelSetPosition(this.panelGetPosition());
+        if (this._timeoutIds['emitPanelPositionChanged']) {
+            this._glib.source_remove(this._timeoutIds['emitPanelPositionChanged']);
+            delete(this._timeoutIds['emitPanelPositionChanged']);
         }
 
-        if (this.isPanelVisible()) {
-            this.panelShow(true, 0);
-        } else {
+        if (this._timeoutIds['emitPanelPositionChanged2']) {
+            this._glib.source_remove(this._timeoutIds['emitPanelPositionChanged2']);
+            delete(this._timeoutIds['emitPanelPositionChanged2']);
+        }
+
+        if (!calledFromChanger) {
+            this.panelSetPosition(this.panelGetPosition(), true);
+        }
+
+        if (!this.isPanelVisible()) {
             let mode = this._panelHideMode ? this._panelHideMode : 0;
             this.panelHide(mode, 0);
+        } else {
+            // resize panel can fix windows going under panel
+            // we may not need it on X11, but it is needed on Wayland
+            // we also need delay after animation
+            // because without delay it many not fix the issue
+            let panelBox = this._main.layoutManager.panelBox;
+            let duration = this._addToAnimationDuration(180);
+            this._timeoutIds['emitPanelPositionChanged']
+            = this._glib.timeout_add(this._glib.PRIORITY_IDLE, duration, () => {
+                delete(this._timeoutIds['emitPanelPositionChanged']);
+                panelBox.hide();
+                this._timeoutIds['emitPanelPositionChangedIn2']
+                = this._glib.timeout_add(this._glib.PRIORITY_IDLE, 20, () => {
+                    delete(this._timeoutIds['emitPanelPositionChangedIn2']);
+                    panelBox.show();
+                    return this._glib.SOURCE_REMOVE;
+                });
+                return this._glib.SOURCE_REMOVE;
+            });
         }
+
+        this._fixLookingGlassPosition();
     }
 
     /**
      * show panel
      *
-     * @param {boolean} force apply show even if it is hidden. defaults to false.
-     * @param {number} animationDuration in miliseconds. defaults to 150 
+     * @param {number} animationDuration in milliseconds. defaults to 150 
      *
      * @returns {void}
      */
-    panelShow(force = false, animationDuration = 150)
+    panelShow(animationDuration = 150)
     {
-        this._panelVisiblity = true;
+        this._panelVisibility = true;
 
         let classname = this._getAPIClassname('no-panel');
 
-        if (!force && !this.UIStyleClassContain(classname)) {
+        if (!this.UIStyleClassContain(classname)) {
             return;
         }
 
@@ -379,7 +448,7 @@ var API = class
             mode: this._clutter.AnimationMode.EASE,
             duration: animationDuration,
             onComplete: () => {
-                // hide and show can fix windows not going under panel
+                // hide and show can fix windows going under panel
                 panelBox.hide();
                 panelBox.show();
                 this._fixLookingGlassPosition();
@@ -411,13 +480,13 @@ var API = class
      *
      * @param {mode} hide mode see PANEL_HIDE_MODE. defaults to hide all
      * @param {boolean} force apply hide even if it is hidden
-     * @param {number} animationDuration in miliseconds. defaults to 150
+     * @param {number} animationDuration in milliseconds. defaults to 150
      *
      * @returns {void}
      */
     panelHide(mode, animationDuration = 150)
     {
-        this._panelVisiblity = false;
+        this._panelVisibility = false;
         this._panelHideMode = mode;
 
         let overview = this._main.overview;
@@ -437,7 +506,7 @@ var API = class
             mode: this._clutter.AnimationMode.EASE,
             duration: animationDuration,
             onComplete: () => {
-                // hide and show can fix windows not going under panel
+                // hide and show can fix windows going under panel
                 panelBox.hide();
                 panelBox.show();
                 this._fixLookingGlassPosition();
@@ -455,12 +524,12 @@ var API = class
             delete(this._overviewHidingSignal);
         }
 
-        let appMenuOriginalVisiblity;
+        let appMenuOriginalVisibility;
 
         if (mode === PANEL_HIDE_MODE.DESKTOP) {
             if (!this._overviewShowingSignal) {
                 this._overviewShowingSignal = overview.connect('showing', () => {
-                    appMenuOriginalVisiblity = this.isAppMenuVisible(); 
+                    appMenuOriginalVisibility = this.isAppMenuVisible(); 
                     this.appMenuHide();
                     panelBox.ease({
                         translation_y: 0,
@@ -476,7 +545,7 @@ var API = class
                         mode: this._clutter.AnimationMode.EASE,
                         duration: 250,
                         onComplete: () => {
-                            if (appMenuOriginalVisiblity) {
+                            if (appMenuOriginalVisibility) {
                                 this.appMenuShow();
                             } else {
                                 this.appMenuHide();
@@ -514,11 +583,11 @@ var API = class
      */
     isPanelVisible()
     {
-        if (this._panelVisiblity === undefined) {
+        if (this._panelVisibility === undefined) {
             return true;
         }
 
-        return this._panelVisiblity;
+        return this._panelVisibility;
     }
 
     /**
@@ -528,7 +597,7 @@ var API = class
      */
     isDashVisible()
     {
-        return this._dashVisiblity === undefined || this._dashVisiblity;
+        return this._dashVisibility === undefined || this._dashVisibility;
     }
 
     /**
@@ -542,7 +611,7 @@ var API = class
             return;
         }
 
-        this._dashVisiblity = true;
+        this._dashVisibility = true;
 
         this._main.overview.dash.show();
 
@@ -566,7 +635,7 @@ var API = class
             return;
         }
 
-        this._dashVisiblity = false;
+        this._dashVisibility = false;
 
         this._main.overview.dash.hide();
 
@@ -671,7 +740,7 @@ var API = class
      * show search
      *
      * @param {boolean} fake true means it just needs to do the job but
-     *   don't need to change the search visiblity status
+     *   don't need to change the search visibility status
      *
      * @returns {void}
      */
@@ -692,13 +761,13 @@ var API = class
             height: searchEntry.height,
             opacity: 255,
             mode: this._clutter.AnimationMode.EASE,
-            duration: 150,
-            onComplete: () => { 
+            duration: 110,
+            onComplete: () => {
                 searchEntryParent.height = -1;
                 searchEntry.ease({
                     opacity: 255,
                     mode: this._clutter.AnimationMode.EASE,
-                    duration: 150,
+                    duration: 700,
                 });
             },
         });
@@ -712,7 +781,7 @@ var API = class
      * hide search
      *
      * @param {boolean} fake true means it just needs to do the job
-     *   but don't need to change the search visiblity status
+     *   but don't need to change the search visibility status
      *
      * @returns {void}
      */
@@ -727,14 +796,13 @@ var API = class
             opacity: 0,
             mode: this._clutter.AnimationMode.EASE,
             duration: 50,
-            onComplete: () => {
-                searchEntryParent.ease({
-                    height: 0,
-                    opacity: 0,
-                    mode: this._clutter.AnimationMode.EASE,
-                    duration: 120,
-                });
-            },
+        });
+
+        searchEntryParent.ease({
+            height: 0,
+            opacity: 0,
+            mode: this._clutter.AnimationMode.EASE,
+            duration: 120,
         });
 
         if (!fake) {
@@ -827,7 +895,6 @@ var API = class
         let signalName = (bySearchController) ? 'notify::search-active' : 'page-changed';
 
         this._searchActiveSignal = controller.connect(signalName, () => {
-
             if (this._searchEntryVisibility) {
                 return;
             }
@@ -854,11 +921,11 @@ var API = class
      */
     OSDEnable()
     {
-        if (!this._originals['osdWindowManager']) {
+        if (!this._originals['osdWindowManagerShow']) {
             return;
         }
 
-        this._main.osdWindowManager.show = this._originals['osdWindowManager'];
+        this._main.osdWindowManager.show = this._originals['osdWindowManagerShow'];
     }
 
     /**
@@ -868,8 +935,8 @@ var API = class
      */
     OSDDisable()
     {
-        if (!this._originals['osdWindowManager']) {
-            this._originals['osdWindowManager']
+        if (!this._originals['osdWindowManagerShow']) {
+            this._originals['osdWindowManagerShow']
             = this._main.osdWindowManager.show;
         }
 
@@ -883,12 +950,20 @@ var API = class
      */
     workspacePopupEnable()
     {
-        if (!this._originals['workspaceSwitcherPopup']) {
+        if (this._shellVersion < 42) {
+            if (!this._originals['workspaceSwitcherPopupShow']) {
+                return;
+            }
+            this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype._show
+            = this._originals['workspaceSwitcherPopupShow'];
+        }
+
+        if (!this._originals['workspaceSwitcherPopupDisplay']) {
             return;
         }
 
-        this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype._show
-        = this._originals['workspaceSwitcherPopup'];
+        this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype.display
+        = this._originals['workspaceSwitcherPopupDisplay']
     }
 
     /**
@@ -898,12 +973,22 @@ var API = class
      */
     workspacePopupDisable()
     {
-        if (!this._originals['workspaceSwitcherPopup']) {
-            this._originals['workspaceSwitcherPopup']
-            = this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype._show;
+        if (this._shellVersion < 42) {
+            if (!this._originals['workspaceSwitcherPopupShow']) {
+                this._originals['workspaceSwitcherPopupShow']
+                = this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype._show;
+            }
+            this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype._show = () => {
+               return false;
+            };
         }
 
-        this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype._show = () => {
+        if (!this._originals['workspaceSwitcherPopupDisplay']) {
+            this._originals['workspaceSwitcherPopupDisplay']
+            = this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype.display;
+        }
+
+        this._workspaceSwitcherPopup.WorkspaceSwitcherPopup.prototype.display = (index) => {
            return false;
         };
     }
@@ -988,6 +1073,24 @@ var API = class
     }
 
     /**
+     * get Secondary Monitor Display
+     *
+     * @returns {ui.WorkspacesView.SecondaryMonitorDisplay}
+     */
+    _getSecondaryMonitorDisplay()
+    {
+        if (this._shellVersion < 40) {
+            return null;
+        }
+
+        // for some reason the first time we get the value it returns null in 42
+        // but it returns the correct value in second get
+        this._workspacesView.SecondaryMonitorDisplay;
+
+        return this._workspacesView.SecondaryMonitorDisplay;
+    }
+
+    /**
      * set workspace switcher to its default size
      *
      * @returns {void}
@@ -1009,8 +1112,8 @@ var API = class
         }
 
         if (this._originals['smd_getThumbnailsHeight'] !== undefined) {
-            let smd = this._workspacesView.SecondaryMonitorDisplay;
-            smd._getThumbnailsHeight = this._originals['smd_getThumbnailsHeight'];
+            let smd = this._getSecondaryMonitorDisplay();
+            smd.prototype._getThumbnailsHeight = this._originals['smd_getThumbnailsHeight'];
         }
 
         this._workspaceSwitcherLastSize = size;
@@ -1044,11 +1147,11 @@ var API = class
             // we are overriding the _getThumbnailsHeight() here with the same
             // function as original but we change the MAX_THUMBNAIL_SCALE to our
             // custom size.
-            // we do this because MAX_THUMBNAIL_SCALE is const and cannot be cahnged
-            let smd = this._workspacesView.SecondaryMonitorDisplay;
+            // we do this because MAX_THUMBNAIL_SCALE is const and cannot be changed
+            let smd = this._getSecondaryMonitorDisplay();
 
             if (this._originals['smd_getThumbnailsHeight'] === undefined) {
-                this._originals['smd_getThumbnailsHeight'] = smd._getThumbnailsHeight;
+                this._originals['smd_getThumbnailsHeight'] = smd.prototype._getThumbnailsHeight;
             }
 
             smd.prototype._getThumbnailsHeight = function(box) {
@@ -1072,7 +1175,7 @@ var API = class
     }
 
     /**
-     * toggle overview visiblity
+     * toggle overview visibility
      *
      * @returns {void}
      */
@@ -1402,13 +1505,19 @@ var API = class
      * move panel position
      *
      * @param {number} position see PANEL_POSITION
+     * @param {boolean} force allow to set even when the current position
+     *   is the same
      *
      * @returns {void}
      */
-    panelSetPosition(position)
+    panelSetPosition(position, force = false)
     {
         let monitorInfo = this.monitorGetInfo();
         let panelBox = this._main.layoutManager.panelBox;
+
+        if (!force && position === this.panelGetPosition()) {
+            return;
+        }
 
         if (position === PANEL_POSITION.TOP) {
             this._panelPosition = PANEL_POSITION.TOP;
@@ -1439,7 +1548,7 @@ var API = class
         if (!this._workareasChangedSignal) {
             this._workareasChangedSignal
             = global.display.connect('workareas-changed', () => {
-                this.panelSetPosition('bottom');
+                this.panelSetPosition(PANEL_POSITION.BOTTOM, true);
             });
         }
 
@@ -1459,18 +1568,24 @@ var API = class
             this._originals['lookingGlassResize'] = lookingGlassProto._resize;
         }
 
-        if (this._panelPosition === PANEL_POSITION.TOP && this.isPanelVisible()) {
+        if (this.panelGetPosition() === PANEL_POSITION.TOP && this.isPanelVisible()) {
 
             lookingGlassProto._resize = this._originals['lookingGlassResize'];
             delete(lookingGlassProto._oldResize);
             delete(this._originals['lookingGlassResize']);
+            if (this._main.lookingGlass) {
+                this._main.lookingGlass._resize();
+            }
 
             return;
         }
 
         const Main = this._main;
 
-        lookingGlassProto._oldResize = this._originals['lookingGlassResize'];
+        if (lookingGlassProto._oldResize === undefined) {
+            lookingGlassProto._oldResize = this._originals['lookingGlassResize'];
+        }
+
         lookingGlassProto._resize = function () {
             let panelHeight = Main.layoutManager.panelBox.height;
             this._oldResize();
@@ -1508,7 +1623,7 @@ var API = class
     }
 
     /**
-     * enable panel notifiction icon
+     * enable panel notification icon
      *
      * @returns {void}
      */
@@ -1518,7 +1633,7 @@ var API = class
     }
 
     /**
-     * disable panel notifiction icon
+     * disable panel notification icon
      *
      * @returns {void}
      */
@@ -1664,7 +1779,7 @@ var API = class
      *
      * @returns {void}
      */
-    enablenAimationsSetDefault()
+    enableAnimationsSetDefault()
     {
         if (this._originals['enableAnimations'] === undefined) {
             return;
@@ -1682,7 +1797,7 @@ var API = class
      *
      * @returns {void}
      */
-    enablenAimationsSet(status)
+    enableAnimationsSet(status)
     {
         if (this._originals['enableAnimations'] ===  undefined) {
             this._originals['enableAnimations']
@@ -1702,12 +1817,12 @@ var API = class
      *
      * @returns {void}
      */
-    ativitiesButtonAddIcon(type, icon, monochrome, holdLabel)
+    activitiesButtonAddIcon(type, icon, monochrome, holdLabel)
     {
-        let iconSize = this._panel.PANEL_ICON_SIZE - this._panel.APP_MENU_ICON_MARGIN;
+        let iconSize = this.panelIconGetSize() - this._panel.APP_MENU_ICON_MARGIN;
         let activities = this._main.panel.statusArea['activities'];
 
-        this.ativitiesButtonRemoveIcon();
+        this.activitiesButtonRemoveIcon();
 
         if (!this._activitiesBtn) { 
             this._activitiesBtn = {};
@@ -1780,7 +1895,7 @@ var API = class
      *
      * @returns {void}
      */
-    ativitiesButtonRemoveIcon()
+    activitiesButtonRemoveIcon()
     {
         let activities = this._main.panel.statusArea['activities'];
 
@@ -1806,6 +1921,28 @@ var API = class
         }
 
         this.UIStyleClassRemove(this._getAPIClassname('activities-button-no-label'));
+    }
+
+    /**
+     * set activities button icon size
+     *
+     * @param {number} size 1-60
+     *
+     * @returns {void}
+     */
+    _activitiesButtonIconSetSize(size)
+    {
+        if (size < 1 || size > 60) {
+            return;
+        }
+
+        let activities = this._main.panel.statusArea['activities'];
+
+        if (!this._activitiesBtn || !this._activitiesBtn.icon) {
+            return;
+        }
+        
+        this._activitiesBtn.icon.icon_size = size - this._panel.APP_MENU_ICON_MARGIN;
     }
 
     /**
@@ -2015,10 +2152,24 @@ var API = class
     notificationBannerPositionSet(pos)
     {
         let messageTray = this._main.messageTray;
+        let bannerBin = messageTray._bannerBin;
 
-        if (this._originals['bannerAlignment'] === undefined) {
-            this._originals['bannerAlignment'] = messageTray.bannerAlignment;
+        if (this._originals['bannerAlignmentX'] === undefined) {
+            this._originals['bannerAlignmentX'] = messageTray.bannerAlignment;
         }
+
+        if (this._originals['bannerAlignmentY'] === undefined) {
+            this._originals['bannerAlignmentY'] = bannerBin.get_y_align();
+        }
+
+        if (this._originals['hideNotification'] === undefined) {
+            this._originals['hideNotification'] = messageTray._hideNotification;
+        }
+
+        // TOP
+        messageTray._hideNotification = this._originals['hideNotification'];
+
+        bannerBin.set_y_align(this._clutter.ActorAlign.START);
 
         if (pos === NOTIFICATION_BANNER_POSITION.TOP_START) {
             messageTray.bannerAlignment = this._clutter.ActorAlign.START;
@@ -2034,6 +2185,78 @@ var API = class
             messageTray.bannerAlignment = this._clutter.ActorAlign.CENTER;
             return;
         }
+
+        // BOTTOM
+
+        // >>
+        // This block is going to fix the animation when the notification is
+        // in bottom area
+        // this is the same function from (ui.messageTray.messageTray._hideNotification)
+        // with clutter animation mode set to EASE.
+        // because the EASE_OUT_BACK (original code) causes glitch when
+        // the tray is on bottom 
+        const State = this._messageTray.State;
+        const ANIMATION_TIME = this._messageTray.ANIMATION_TIME;
+        const Clutter = this._clutter;
+
+        messageTray._hideNotification = function (animate) {
+            this._notificationFocusGrabber.ungrabFocus();
+
+            if (this._bannerClickedId) {
+                this._banner.disconnect(this._bannerClickedId);
+                this._bannerClickedId = 0;
+            }
+            if (this._bannerUnfocusedId) {
+                this._banner.disconnect(this._bannerUnfocusedId);
+                this._bannerUnfocusedId = 0;
+            }
+
+            this._resetNotificationLeftTimeout();
+            this._bannerBin.remove_all_transitions();
+
+            if (animate) {
+                this._notificationState = State.HIDING;
+                this._bannerBin.ease({
+                    opacity: 0,
+                    duration: ANIMATION_TIME,
+                    mode: Clutter.AnimationMode.EASE,
+                });
+                this._bannerBin.ease({
+                    opacity: 0,
+                    y: this._bannerBin.height,
+                    duration: ANIMATION_TIME,
+                    mode: Clutter.AnimationMode.EASE,
+                    onComplete: () => {
+                        this._notificationState = State.HIDDEN;
+                        this._hideNotificationCompleted();
+                        this._updateState();
+                    },
+                });
+            } else {
+                this._bannerBin.y = this._bannerBin.height;
+                this._bannerBin.opacity = 0;
+                this._notificationState = State.HIDDEN;
+                this._hideNotificationCompleted();
+            }
+        }
+        // <<
+
+        bannerBin.set_y_align(this._clutter.ActorAlign.END);
+
+        if (pos === NOTIFICATION_BANNER_POSITION.BOTTOM_START) {
+            messageTray.bannerAlignment = this._clutter.ActorAlign.START;
+            return;
+        }
+
+        if (pos === NOTIFICATION_BANNER_POSITION.BOTTOM_END) {
+            messageTray.bannerAlignment = this._clutter.ActorAlign.END;
+            return;
+        }
+
+        if (pos === NOTIFICATION_BANNER_POSITION.BOTTOM_CENTER) {
+            messageTray.bannerAlignment = this._clutter.ActorAlign.CENTER;
+            return;
+        }
     }
 
     /**
@@ -2043,12 +2266,19 @@ var API = class
      */
     notificationBannerPositionSetDefault()
     {
-        if (this._originals['bannerAlignment'] === undefined) {
+        if (this._originals['bannerAlignmentX'] === undefined ||
+            this._originals['bannerAlignmentY'] === undefined ||
+            this._originals['hideNotification'] === undefined
+        ) {
             return;
         }
 
         let messageTray = this._main.messageTray;
-        messageTray.bannerAlignment = this._originals['bannerAlignment'];
+        let bannerBin = messageTray._bannerBin;
+
+        messageTray.bannerAlignment = this._originals['bannerAlignmentX'];
+        bannerBin.set_y_align(this._originals['bannerAlignmentY']);
+        messageTray._hideNotification = this._originals['hideNotification'];
     }
 
     /**
@@ -2448,6 +2678,186 @@ var API = class
         });
 
         this._isDoubleSuperToAppGrid = false;
+    }
+
+    /**
+     * set default OSD position
+     *
+     * @returns {void}
+     */
+    osdSetDefaultPosition()
+    {
+        if (!this._originals['osdWindowRelayout']) {
+            return;
+        }
+
+        let osdWindowProto = this._osdWindow.OsdWindow.prototype;
+
+        osdWindowProto._relayout = this._originals['osdWindowRelayout'];
+
+        delete(osdWindowProto._oldRelayout);
+        delete(this._originals['osdWindowRelayout']);
+
+        let osdWindows = this._main.osdWindowManager._osdWindows;
+        osdWindows.forEach(osdWindow => {
+            osdWindow._box.translation_x = 0;
+            osdWindow._relayout();
+        });
+    }
+
+    /**
+     * set OSD position
+     *
+     * @param int x percentage 0 - 100
+     * @param int y percentage 0 - 100
+     *
+     * @returns {void}
+     */
+    osdSetPosition(x, y)
+    {
+        let osdWindowProto = this._osdWindow.OsdWindow.prototype;
+
+        if (!this._originals['osdWindowRelayout']) {
+            this._originals['osdWindowRelayout'] = osdWindowProto._relayout;
+        }
+
+        if (osdWindowProto._oldRelayout === undefined) {
+            osdWindowProto._oldRelayout = this._originals['osdWindowRelayout'];
+        }
+
+        const Main = this._main;
+
+        osdWindowProto._relayout = function () {
+            this._oldRelayout();
+
+            let monitor = Main.layoutManager.monitors[this._monitorIndex];
+            if (!monitor) {
+                return;
+            }
+
+            let negativeX = (x < 50) ? -1 : 1;
+            let negativeY = (y < 50) ? -1 : 1;
+
+            // TODO we need to get actual size of OSD and then subtract to half of that
+            // so we have more accurate translation for both x and y
+            this._box.translation_x
+            = Math.round((monitor.width * x) / 100 - (monitor.width / 2)) * negativeX;
+            this._box.translation_y
+            = Math.round((monitor.height * y) / 100 - (monitor.height / 2)) * negativeY;
+        };
+
+        let osdWindows = this._main.osdWindowManager._osdWindows;
+        osdWindows.forEach(osdWindow => {
+            osdWindow._relayout();
+        });
+    }
+
+    /**
+     * show weather in date menu
+     *
+     * @returns {void}
+     */
+    weatherShow()
+    {
+        this.UIStyleClassRemove(this._getAPIClassname('no-weather'));
+    }
+
+    /**
+     * hide weather in date menu
+     *
+     * @returns {void}
+     */
+    weatherHide()
+    {
+        this.UIStyleClassAdd(this._getAPIClassname('no-weather'));
+    }
+
+    /**
+     * show world clocks in date menu
+     *
+     * @returns {void}
+     */
+    worldClocksShow()
+    {
+        this.UIStyleClassRemove(this._getAPIClassname('no-world-clocks'));
+    }
+
+    /**
+     * hide world clocks in date menu
+     *
+     * @returns {void}
+     */
+    worldClocksHide()
+    {
+        this.UIStyleClassAdd(this._getAPIClassname('no-world-clocks'));
+    }
+
+    /**
+     * set default panel icon size
+     *
+     * @returns {void}
+     */
+    panelIconSetDefaultSize()
+    {
+        if (this._panelIconSize === undefined || !this._originals['panelIconSize']) {
+            return;
+        }
+
+        let classnameStarter = this._getAPIClassname('panel-icon-size');
+        this.UIStyleClassRemove(classnameStarter + this._panelIconSize);
+        this._emitRefreshStyles();
+
+        let defaultSize = this._originals['panelIconSize'];
+        this._panel.PANEL_ICON_SIZE = defaultSize;
+        this._main.panel.statusArea['dateMenu']._indicator.set_icon_size(defaultSize);
+        this._main.panel.statusArea['appMenu']._onIconThemeChanged();
+        this._activitiesButtonIconSetSize(defaultSize);
+
+        delete(this._panelIconSize);
+    }
+
+    /**
+     * set panel icon size
+     *
+     * @param {number} size 1-60
+     *
+     * @returns {void}
+     */
+    panelIconSetSize(size)
+    {
+        if (size < 1 || size > 60) {
+            return;
+        }
+
+        if (!this._originals['panelIconSize']) {
+            this._originals['panelIconSize'] = this._panel.PANEL_ICON_SIZE;
+        }
+
+        let classnameStarter = this._getAPIClassname('panel-icon-size');
+        this.UIStyleClassRemove(classnameStarter + this.panelIconGetSize());
+        this.UIStyleClassAdd(classnameStarter + size);
+        this._emitRefreshStyles();
+
+        this._panel.PANEL_ICON_SIZE = size;
+        this._main.panel.statusArea['dateMenu']._indicator.set_icon_size(size);
+        this._main.panel.statusArea['appMenu']._onIconThemeChanged();
+        this._activitiesButtonIconSetSize(size);
+
+        this._panelIconSize = size;
+    }
+
+    /**
+     * get panel icon size
+     *
+     * @returns {void}
+     */
+    panelIconGetSize()
+    {
+        if (this._panelIconSize !== undefined) {
+            return this._panelIconSize;
+        }
+
+        return this._panel.PANEL_ICON_SIZE;
     }
 }